Fix GiveEgg original bug
GiveEgg scripts logic is backwards, so if you receive an egg of a
Pokémon that you have seen but not caught, the game will mark it as
caught but not seen, deleting its entry from the Pokédex until you see
the mon again.

This makes the GiveEgg work correctly.